# contract/openapi.yaml
openapi: 3.1.0
info:
  title: MyBooking API
  version: 0.0.0
...
paths:
...
  /api/v1/bookings/{bookingId}/confirm:
    post:
      operationId: confirmBooking
      description: Только организатор; Origin и CSRF. Проверяются срок, состояние и занятость обоих участников.
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: string
        - name: X-CSRF-Token
          in: header
          required: true
          schema:
            type: string
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpertBookingDetail'
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: Access is forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: The request conflicts with the current state of the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
        - ApiKeyAuth: []
...
components:
  schemas:
...
    ApiError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Публичный код без раскрытия чужих сведений.
        message:
          type: string
          description: Безопасное описание на русском языке.
...
    BookingDetail:
      type: object
      required:
        - id
        - status
        - startAt
        - endAt
        - topic
        - guestName
        - guestTimezone
        - expertName
        - expertPublicId
        - completed
      properties:
        id:
          type: string
        status:
          type: string
          enum:
            - pending
            - confirmed
            - rejected
            - withdrawn
            - expired
            - cancelled
        startAt:
          $ref: '#/components/schemas/OffsetDateTime'
        endAt:
          $ref: '#/components/schemas/OffsetDateTime'
        topic:
          type: string
        description:
          type: string
        guestName:
          type: string
        guestTimezone:
          type: string
        expertName:
          type: string
        expertPublicId:
          type: string
        closedReason:
          type: string
          enum:
            - manual
            - conflict
            - schedule_changed
            - deadline
        completed:
          type: boolean
...
    ExpertBookingDetail:
      type: object
      required:
        - guestEmail
      properties:
        guestEmail:
          type: string
          format: email
      allOf:
        - $ref: '#/components/schemas/BookingDetail'
...
    OffsetDateTime:
      type: string
      format: date-time
      description: Момент ISO 8601 с явным смещением; сервер определяет однозначный UTC-момент.
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: cookie
      name: mybooking_session
    BearerAuth:
      type: http
      scheme: Bearer